Add babel plugin to improve SSR and create Layout.style file
