+++ package.json
@@ -26,6 +26,7 @@
     "@styled-system/theme-get": "^5.1.2",
     "@testing-library/user-event": "^12.8.3",
     "axios": "^0.21.1",
+    "babel-plugin-styled-components": "^1.12.0",
     "i18next": "^19.9.0",
     "next": "10.0.7",
     "react": "17.0.1",
