+++ public/locales/en/HomeForm.json
@@ -1,4 +1,3 @@
 {
-  "follow": "follow",
   "search": "search"
 }
