+++ src/common/layouts/Layout.tsx
@@ -1,26 +1,13 @@
-import { themeGet } from '@styled-system/theme-get';
 import { ReactElement } from 'react';
-import styled from 'styled-components';
 
-import { Layout as UiCommonsLayout } from '@olist/ui-commons/dist/ui-commons.cjs';
 import { Flex } from '@olist/united';
 
 import Footer from '~common/components/Footer';
+import { Container } from '~common/layouts/Layout.style';
 
 interface Layout {
   children: JSX.Element[] | JSX.Element;
 }
-
-export const Container = styled(UiCommonsLayout.Container).attrs({
-  py: 2,
-  px: 3,
-})`
-  background: ${themeGet('colors.foundation.shadeWhite')};
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-  max-width: 1440px;
-`;
 
 const Layout = ({ children }: Layout): ReactElement => {
   return (
