+++ src/pages/_document.tsx
@@ -1,9 +1,36 @@
 import Document, { Html, Head, Main, NextScript } from 'next/document';
 
+import { ReactElement } from 'react';
+import { ServerStyleSheet } from 'styled-components';
+
 // eslint-disable-next-line no-use-before-define
-import React, { ReactElement } from 'react';
 
 class AppDocument extends Document {
+  static async getInitialProps(ctx) {
+    const sheet = new ServerStyleSheet();
+    const originalRenderPage = ctx.renderPage;
+
+    try {
+      ctx.renderPage = () =>
+        originalRenderPage({
+          enhanceApp: (App) => (props) => sheet.collectStyles(<App {...props} />),
+        });
+
+      const initialProps = await Document.getInitialProps(ctx);
+      return {
+        ...initialProps,
+        styles: (
+          <>
+            {initialProps.styles}
+            {sheet.getStyleElement()}
+          </>
+        ),
+      };
+    } finally {
+      sheet.seal();
+    }
+  }
+
   render(): ReactElement {
     return (
       <Html>
